requirements.txt
Flask
Flask-SQLAlchemy
flask-script
requests
marshmallow